macosx build.sh: Automatically detect HOST_ARCH
Do not hardcode HOST_ARCH to x86_64, but read out the actual arch from
the host system. This allows to always build a native build by default,
irregardless if you start the script on an intel or arm mac.

You can still overwrite the value using -a command line, in order to
start a cross compilation.