webvtt: do not hardcode output file name
Instead of hardcoding the output file name in the
.l file, specify it on the command line instead.

In theory it would not be necessary to specify it at all as lex.yy.c
is the default, however overwriting the prefix will change this, so
we have to force it back to lex.yy.c so that ylwrap still works as
expected and can find the output file.

We do not hardcode it in the file anymore so that for Meson it can
properly specify the --outfile option and have it not overwritten by
the per-file option.