win32: drawable: replace the internal HWND with a hook on the parent Window
Every time the parent receives a message, we check if the size has not changed.

The window hook only works if we have a DLL. This is the case in regular windows
builds. It won't work if it's very built statically.

We do not use an internal HWND anymore that was not always getting the WM_SIZE
events.