input_decoder: guard vlc_aout_stream_Play()
Like it's already done for Video in 7dcc370d120266ca552433b0b419f1e65f194f5c

Fixes vlc_aout_stream functions that could be called in parallel (Play +
Drain in //), leading to assert, crashes, or undefined behavior in aout
implementations.

It was not done for audio because vlc_aout_stream_Play() was blocking.

Main aout modules are not blocking anymore, we still need to fix the
less important ones.

Refs #27508