control: dbus: emit CanPlay consistently
When VLC was launched from a file, the handling for playlist
change events was not sufficient as in case of starting with
a file, we could miss the event and never report CanPlay.

When playing a file from a service discovery, like our Lua
web parser scripts, this logic was flawed too, as the playlist
would always be "empty" even though we are clearly currently
paused on an input. To fix that, check if CanPlay disagrees
with reality when we get an input change and emit CanPlay.

There is no check added in the item append/delete sections,
so currently once an item from an SD is played, even when
the user has switched back to the main playlist, the status
would still be CanPlay until and item is added or removed
in the "main" playlist. However this exactly matches how the
Qt interface behaves, so I've left this behavior as-is for
consistency with the interface.

Fix #21419
Fix #22785