contrib: wix: add the Wix tool as a windows prebuilt bin package
Rather than relying on it being installed somewhere on wine we can use our own
and control which version is used.

It's always added to the windows builds even though it may not be used in the
end. But the package is not big and doesn't require any building.

It requires dotnet 4.0 being installed in Wine.

While WIX is open source, the wix 3.5 we are using cannot realistically be
built in our contribs as it's hardcoded to build with VS 2008. Even the most
recent builds of WIX still don't build in Linux. We already use a prebuilt
version in Docker images, except it's a .msi file that is installed via wine.