VLCOpenGLES2VideoView: remove EAGL context in close
Fix a crash where the destructor of the VLCOpenGLES2VideoView
(.cxx_destruct) is calling the destructor of the EAGL context, which
leads into unknown stacktraces in the EAGLContext_setDebugLabel symbol.

The setup of the crash needs using CVPixelBuffer backed by an IOSurface,
so typically a filter using the VLCCVOpenGLProvider offscreen OpenGL
implementation, which seems to imply some live-binding is still in use
if we don't close the context soon enough.