Revert "contrib: qt: don't patch create_libtool value"
This reverts commit cee6cee6aacd9ac186e91c9fbfa2e7c8e9fc6033.

Unfortunately, removing create_libtool from CONFIG in the toolchain file
gets overridden when it is added to CONFIG again in qt_module.prf.

Qt itself adds it to CONFIG for "!lib_bundle:unix", indeed in the
Windows contrib archive, libQt5Bootstrap.la can be spotted (as this is
for the native build only). Also, when building for Debian, all Qt
modules will install .la files, like libQt5Core.la.

As qt_module.prf is the only place in the whole of Qt where this is
added to CONFIG, it should be sufficient to patch it here.