transcode: handle first-PCR issues
Previously, we fast-forwarded the PCR values no matter what. This used
to be handy but caused plausible invalid PCR as the fast-forwarded first
PCRs values can't be checked against future encoded DTS.

This patch approach the problem by synthesizing the first PCR as
`VLC_TICK_0` to ensure no following DTS is lower and still giving a
starting point of the stream to the next modules. We also avoid
forwarding PCR values before any input is signaled to the `pcr_sync`
utility, to avoid the exact same issue talked above.