meson: sapi: autodetect support on Windows
We cannot assume that the required headers and type are always present
when building for Windows. On mingw-llvm, the headers are present, but
restricted to desktop usage only. Closely mirror configure.ac, which
also performs these checks.

The difference is that we now only perform these checks when building
for Windows: configure.ac performed the checks unconditionally.