vlcrs-macros: meson: add definition for crate
Co-authored-by: Alexandre Janniaux <ajanni@videolabs.io>

Most of the code was written by Loïc and extracted from his original
merge request, but those snippets were not merged upstream. The code was
extracted to a separate meson build definition.

The added extension addition allows ensuring the built artifact matches
with what cargo will build. Since it's a proc-macro crate, it will build
a plugin for the compiler, which is a dynamic library of the building
platform.

The plugin itself is not used currently, but this allows checking that
the proc_macro builds correctly.