medialibrary: Thumbnailer: fix double-free
The picture was referenced on the ThumbnailerCtx object, which release
it whenever one is not nullptr. However, after the changes in !3088,
commit 2ea2c7eb19122930877d42e981a6ef97e90a1f6e, the release policy for
the picture being exported has been fixed to ensure the picture is
consumed whatever the conversion case.

Ensure the picture won't be released twice by swapping it from the
ThumbnailerCtx object before exporting.

Fix #27749