transcode: pcr_sync: switch assert to runtime error
This assertion was too severe and deserve actual runtime handling.

The `pcr_helper` should just fail in case of inconsistent
input/output. This can happen in various case such as bogus encoder or
simply internal error in the `pcr_helper` code.

This patch is making sure the `pcr_helper` user can disable it in case
of failure.

Refs #27730