audio_output: android: move AudioFormat vtable
Move the AudioFormat vtable to a separate file and convenience library
so as to use the AudioFormat vtable from device.c.

The choice of a convenience library here is to compile the code only
once but be able to use it from different locations, especially since it
needs the vtable filled before using the associated function.

The commit doesn't change audiotrack yet, as the goal is mostly to fix
the dynamic build on Android, and thus duplicates part of the code from
there. Some more changes would be needed to potentially supply the
vtable as parameter and provides what's missing for audiotrack.c to use
this convenience library instead.