transcode: disable pcr forwarding on error
`SignalLeavingFrame` returning an error usually means that either the
decoder outputs an inconsistent stream timestamp-wise or that the PCR
sync algorithm had a mistake in its implementation.

Both case shouldn't ignore the error, from that point, the PCR
output should be considered flawed and PCR output from transcode
should be stopped with a proper error message.

Refs #27730