test: lua: check that extension are correctly probed
The test included in this commit checks that lua will correctly probe
extension scripts, but without ensuring that it correctly follows the
data home. It only check that lua creates the correct access to load
the files.

In particular, the test fails to pass if the condition for loading .vle
files is not correct.