sout: display: add clock support
Create a local clock context using the new `sout_Clock` API. This set
of clocks will be passed to the input decoder to have a synchronization
context local to the display module.

Currently, no master clock is selected for the sake of simplicity,
implicitly falling back to monotonic master.

Fixes #23220