image: ensure picture is released from client
Ensure the client calling the export to block_t is also responsible for
releasing the picture it exports.

Previous commits were adding the release to match behaviour of the
converter before refactoring everything, but it was a mistake and led to
issues with other parts of the pipeline, in particular the transcoder,
and some clients were more adapted to the converter path + hold of
original rather than the non-conversion path, which would then have been
simpler to do from the start.

This MR adds the necessary hold and release to remove those from the
encoders and the clients, and is tested against the unmerged
libvlc/media.c test for thumbnailing too.

Revert "png: fix picture release on encoder"
This reverts commit d4c93a29a223c6d68e3336f5312a03644896e09e.

Revert "jpeg: fix picture release on encoder"
This reverts commit cccc085e30d6ca4862f43bf93372e2fc156c1778.

Revert "avcodec: encoder: fix picture release"
This reverts commit 5a145f34fa463ac458e65e9f1b2a2e2595c58494.

Revert "codec: vpx: fix picture release on encoder"
This reverts commit 5dde7db39c1c9fd1f72b28dc1786798300b6106c.

Revert "image: fix picture leak in case of error"
This reverts commit 1bf09fe8d6ca2b7d90d48b29f8d0f2164a84d382.

Revert "misc: thumbnailer: release picture when interrupted"
This reverts commit 004b9900f837147d49895e96f86ae2c1ef4c3dea.