lua: add ext parameter
Every controls are modifying an extension and require the extension to
be non-null as first parameter. Instead of forwarding the extension to
the va_list, extract it as first parameters unconditionnally.

This prepares some work to replace the pf_control callback by dedicated
operations and start reworking some code style along the way.