lua: extension: listen to player
Use a player listener per extension to signal input changes instead of
relying on the UI to supply any new input change. This is using one
listener per extension so that we can remove it when deleting the
extension and be signalled of the first input when attaching the
listener the first time.

Fixes #22585