lua: extension: refactor error path to goto
Split the normal path and error path and move the release behind the
error label to remove the handling of the `b_ok` state in the normal
path.

Since returning an error only makes the processing forward to the next
extension to be run, and doesn't fail completely, the label is named
"discard" to signal it will only discard the current extension.