vlc_es: add HDR10+ dynamic metadata
Following the precedent set by the DoVi metadata, we encode this as an
ancillary attached directly to the frame, since it contains
frame-dynamic data.

We make the decision to follow ATSC A/341 Amendment 2094-40 instead of
the original SMPTE ST2094-40:2016 specification, because the latter
includes a lot of extra metadata fields which the former intentionally
leaves unused, and which are also unused in practice (both in
implementation and in files). We can easily add them later should a need
arise.

Also, again following the precedent set by `vlc_video_dovi_metadata_t`,
we encode AVRationals directly as `float`. This saves us from having to
do another unnecessary conversion step in the actual usage site (e.g.
the libplacebo vout) by allowing the use of `av_q2d` directly.