player: timer: handle seek state
Keep the last requested seek position/time, that will be used for
successive jumps. It will be more precise than using the demux
GET_TIME/GET_POSITION.

Send 'on_seek' callback with a valid point when starting a new seek.
This callback will be called with a NULL point once all seek requests
are processed.

UI have now the choice to update their time/position to the seeked
value. This will fix the UI not updated when the seek is long to
process.

Refs #27383
Refs #28482