cpu: remove the old SSE register tests
Since CAN_COMPILE_SSE is no longer defined, this has been silently
disabled. Besides:
- This was always compiled out on Windows and OS/2.
- This is no longer linked in on Linux since the more robust
  /proc/cpuinfo checks were introduced.
- This is not used on MacOS at least since X86-32 support was dropped.

The point of this test was to verify that the OS scheduler supports
the SSE register bank. According to the introduction commit 22 years
ago (2788bc6a4e38026492aa9fb1f2b3b01cad61207d), the problem affected
Linux 2.2.x, which is long since unsupported.

In the end, we still test the CPU feature support via CPUID. Remove
this test only implicitly assumes that the OS kernel is not broken.

For what it is worth, this "fixes" the use of _exit() inside library
code which was flagged by some linting tools.