plugins: rename __PLUGIN__ to VLC_DYNAMIC_PLUGIN
We should not use names stating with underscore for our internal code.
This is almost the same name used in the makefile for more coherence.

HAVE_DYNAMIC_PLUGINS already exists in the core. We keep a different
name to separate the 2 usages. For now they mean exactly the same thing.