mediacodec: release timestamp_fifo with video context
`CleanInputVideo()` is linked to the input side of the decoder
(.pf_decode) but `video.timestamp_fifo` can be used from the output side
(threads from mediacodec) and is actually also tied to the life cycle of
the video context, as visible in the destructor `CloseDecoder`. Since
the life cycle of the video context is at least the one from the input
side, destroy the timestamp_fifo there.

Fixes a use-after-free when the video context is not released from
CloseDestructor() while mediacodec is still running (not joined yet).

Regression from cdff503ed20f5e58aa35b330bfe9fd811c2b76a0 and
57323ddadbe6579697cc9ecdef0de2fbc6831dac