vout: vsync: continue resample after first frame
Change the vout vsync scheduler behaviour on the first frame to ensure
that moving to the render state always happens after either a null
picture or a future picture has been prepared from the prepare state.

Before this commit, the first frame was displayed regardless of its date
and the date of the next picture.