player: input: change ES delay without controls
When setting delay, audio output flushing and picture drops can happen
and lead to unpleasant result, like stuttering, on both outputs.

Since the delay was set by an input command, demux could send a bit of
data to play before the delay was really applied, leading to this effect
even when starting the playback.

This commit alone doesn't fix the issue, but it ensures that delay is
synchronously set before the start of the input when it's not running
yet, mimicking the other input methods allowing to apply a control
synchronously when the input is not yet running.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>

Refs #27918