media_track: fix tracklist leak when no track found
When count is 0 the list is still allocated. Either we should not allocate
one, we free it locally, or we pass it on as it.
The latter is similar to the code from 40c14fbee48975100aaa92b83a99ac33fa086933
in libvlc_media_tracklist_from_es_array().

Introduced in 057f3ef9e8b6bef1db79c1824daf667f4e777283.