vout: new native display renderer for Apple platforms
This new module provides the ability to render Core Video pixel buffers pictures into an AVSampleBufferDisplayLayer.
Pictures decoded by a software decoder are converted to Core Video pixel buffers and picture decoded by the VideoToolbox decoder are rendered directly.
With this native API, rendering will be able to handle HDR tone-mapping natively without using OpenGL/libplacebo as it will be backed by Metal.
It also opens the ability to handle Picture in Picture mode on supported Darwin platforms with upcoming additional work.

It currently have the limitation to not support spherical video as it doesn't seems to be feasible at this time.

This module has the highest priority and will be the first vout display to be opened.