dav1d: buffer more pictures when super resolution is used
When decoding AV1 bitstreams using super resolution, we end up deadlocking
in the decoder pool because we don't have engouh pictures.
We need to add more pictures to the pool so that the decoder works properly.
It seems to need only one extra when using a single thread and 2 when using
more that one thread.

Ref. #28002