win32/common: handle the sensor init/close in the HWND handler
When we create the window we know if it's for a projection or not.
That's all we need to initialize the sensors.

The code is disabled for display module not using the sensor (no projection).
The code is in C++ so we don't want to pull it when we don't need to.