input: decoder: continue DecoderThread on flush
If the decoder thread is flushing, we don't want to stop right now but
instead we'll finish the calls to pf_flush and ensure everything is in
order for this last loop of the decoder thread.

Without this, an input decoder could miss the call to pf_flush and other
output flush before being deleted.