CI: don't block builds when snap build fails
We need to check what is wrong but usually it's the same as Debian builds.
The snap building depending on Ubuntu server is not very stable. So we should
not care too much when it breaks.

(cherry picked from commit 48a6b25031c13d179632d31ecd3c9a72f271e35a)
Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>