extras/tools: use 'command -v' to assert commands
Currently we are using commandname --version to check wheather a
command exists. Hence if a command is not shipped with --version the
check fails. For example busybox implementation of sha512sum fails
to get detected via this method.

Use command -v to check whether a command exists to fix this limitation.