contrib: libplacebo: update to 5.264.1
We need to force the path to find SPIRV otherwise find_library() doesn't
find it even though it's in the contrib (from glslang). find_library()
doesn't use the `libdir` we pass when configuring the code. It seems more
designed to find system libraries. But it's possible to add custom pathes
to find more libraries. This is what the added patch does.