contrib/upnp: disable usage of reuseaddr on iOS/tvOS and fix option
Use of SO_REUSEADDR leads to a busy loop within libupnp and very high
battery usage. This disables the option and actually fixes its
functionality by patching the library. Patch to be upstreamed.

According to App Store analytics, this is the most prominent source of
high CPU load and battery use, so it could explain #27036