android: display: always call render_ts()
From the Android documentation:

"""
if multiple buffers are sent to the Surface to be rendered at the same
VSYNC, the last one will be shown, and the other ones will be dropped.

if the timestamp is not "reasonably close" to the current system time,
the Surface will ignore the timestamp, and display the buffer at the
earliest feasible time. In this mode it will not drop frames.
"""

So, if the picture date is in the past, render_ts() will display it
immediately (unless a future picture with the same VSYNC is pushed just
after).

If the picture date is more than 1 seconds in the future, render_ts()
will behave like render().