qt: optimize RoundImage generation
RoundImage are generated asynchronously, so if a view request again the same
image before it generation ends and reach the cache, a new request is recreated
and the image is re-generated. This patch optimize the image generation by
exposing the request as a shared objects that will be shared amongst the
different image that makes the same request.