mmdevice: move MTA initialization to the worker thread
MTA was checked in the Open function. As this function is called from libvlc
thread, this implies that a user of libvlc should have its own thread in MTA
which may not be possible for various reasons.

* MuteSet, VolumeSet, Play, DeviceSelect, Close callback don't require MTA
* Flush, Pause and Start callbacks still require MTA, but as these functions are
  called from VLC internal threads this is a lesser issue

(cherry-picked from 2fa81b6f15f268e1aaeaf55d93f5bb628714d8cb)