sout: hls: add the playlist architecture
Create the structures and basis for HLS playlist/track management.

This architecture will allow the user to chose which ESes will be muxed
together. An array describing the playlist layout is stored in the HLS
context allowing every ESes to be dispatched to the right playlist in
the `pf_add` callback.

Every ES that does not match the predetermined layout will be assigned
to a new `EXT-X-MEDIA` playlist.