modules: Makefile.am: clean rust artifacts
This commit signal the mostlyclean and other dependant targets that it
should clean the targets generated by cargo.

We use a separate target name and add the dependency to
mostlyclean-local so that we can still extend the known automake target.

We also don't use CLEANFILES so that dependencies files are also removed
when using `make mostlyclean-local-rust` to avoid a corrupted tree where
the dependencies would not be removed but the target files they
reference would not exist anymore.