sout: hls: implement HLS segmentation
To create the HLS segment we refer to the PCR value as a stream time
reference, this allow to expose all the segments at the same time and
handle non-continuous data streams (such as subtitles).

Hence, the muxed output from all tracks is bufferized and split
according to the ideal segment length when a PCR value reaches the
correct timestamp. For now, for the sake of simplicity and the fact that
the main use case of HLS (chromecast renderer) doesn't require it, we
don't ensure video segments starts with an I-Frame. This will
definitely be implemented in a future version as there is already the
mechanism ready for that in the TS muxer.