opengl: expose orientation instead of vflip
Like in the D3D11 implementation, any OpenGL implementation and
especially those coming from the user might ask a specific rendering
orientation to remove the need to rotate the image after it has been
done by the renderer.

In particular, the libvlc output callback allows any orientation to be
supplied by the user, but was currently ignored in release and was
asserting in debug because it didn't support that.

This commit changes the definition and usage of the opengl (offscreen)
providers to expose a format instead. It is now also exposed for
on-screen implementations also.