opengl: filters: forward orientation from gl
Copy the surface orientation property from the vlc_gl_t provider and
store it inside the opengl filters object at initialization. The
parameter is then forwarded only for the last filter into the rendering
metadatas, allowing this filter to enforce the final rendering of the
frame with the correct orientation.