vgl: handle updateCb() failure during Open()
When updateCb() fails, the result was ignored and vgl continued to
Open() as normal in release mode, and was even asserting on debug
before.

This split the Resize() function (which cannot return errors) into
a different ResizeInternal() function with an error return value, to be
called from the Open() instead of the .resize() callback.